+++ src/onegov.feriennet/onegov/feriennet/app.py
@@ -40,7 +40,7 @@ BANNER_TEMPLATE = """
 class FeriennetApp(OrgApp):
 
     #: the version of this application (do not change manually!)
-    version = '1.5.14'
+    version = '1.6.0'
 
     request_class = FeriennetRequest